chores/doc: Restructured the project as a better base for docs and testing
Restructured the project and moved SD and EPUB handling and the UI implementation
in lib, as well as the headers in include the follow the Zephyr example
application and provide a cleaner structure which is better suitable for
generating documentation and testing.

Add API doc.
