+++ .github/workflows/build-zephyr.yaml
@@ -4,7 +4,7 @@
 
 name: Build Zephyr ZEReader Application
 
-on: workflow_dispatch
+on: [push, pull_request, workflow_dispatch]
 
 jobs:
   reuse-license-check:
