+++ CMakeLists.txt
@@ -2,9 +2,7 @@
 #
 # SPDX-License-Identifier: MPL-2.0
 
-# zephyr_syscall_include_directories(include)
+zephyr_syscall_include_directories(include)
+zephyr_include_directories(include)
 
-# zephyr_include_directories(include)
-
-#add_subdirectory(drivers)
-# add_subdirectory(lib)
+add_subdirectory(lib)
