--- app/src/epub/CMakeLists.txt
@@ -1,6 +1,0 @@
-# SPDX-FileCopyrightText: 2025 Anna-Lena Marx <mail@marx.engineer>
-#
-# SPDX-License-Identifier: MPL-2.0
-
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/sd.c)
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/epub.c)
