--- app/src/ui/CMakeLists.txt
@@ -1,9 +1,0 @@
-# SPDX-FileCopyrightText: 2025 Anna-Lena Marx <mail@marx.engineer>
-#
-# SPDX-License-Identifier: MPL-2.0
-
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/context.c)
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/ui.c)
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/ZEReaderLogo.c)
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/zereaderlogomarx.c)
-target_sources(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/notoserif_14.c)
