+++ doc/Doxyfile
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2025 Anna-Lena Marx
+#
+# SPDX-License-Identifier: CC0-1.0
+
 # Doxyfile 1.9.6
 
 # This file describes the settings to be used by the documentation system
