+++ doc/Makefile
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2025 Anna-Lena Marx
+#
+# SPDX-License-Identifier: CC0-1.0
+
 # Minimal makefile for Sphinx documentation
 #
