+++ doc/_doxygen/main.md
@@ -1,3 +1,9 @@
+<!--
+SPDX-FileCopyrightText: 2025 Anna-Lena Marx
+
+SPDX-License-Identifier: MPL-2.0
+-->
+
 # Introduction
 
 This is the Doxygen documentation for [ZEReader].
