+++ doc/docs/debugging.rst
@@ -1,3 +1,7 @@
+.. SPDX-FileCopyrightText: 2025 Anna-Lena Marx
+..
+.. SPDX-License-Identifier: MPL-2.0
+
 Debugging
 =========
