+++ doc/requirements.txt
@@ -1,1 +1,5 @@
+# SPDX-FileCopyrightText: 2025 Anna-Lena Marx
+#
+# SPDX-License-Identifier: CC0-1.0
+
 Sphinx
