+++ include/app/lib/ui/context.h
@@ -7,16 +7,29 @@
 #ifndef _CONTEXT_H_
 #define _CONTEXT_H_
 
+/**
+ * @file
+ * @brief The UI's context.
+ * @defgroup ui_context UI context
+ * @ingroup ui
+ * @{
+ */
+
+/**
+ * @brief UI context definition.
+ */
 typedef enum
 {
-    MENU = 0,
-    READING,
-    SETTINGS
+    MENU = 0,   /**< Menu context. */
+    READING,    /**< Reading context. */
+    SETTINGS    /**< Settings context. */
 
 } context_t;
 
 // When changing the enum, also change the corresponding strings
 // in the .c file!
 extern const char *context_strings[];
+
+/** @} */
 
 #endif
