+++ include/app/lib/ui/logo/ZEReaderLogo.c
