+++ include/app/lib/ui/logo/zereaderlogomarx.h
