+++ lib/CMakeLists.txt
@@ -2,6 +2,6 @@
 #
 # SPDX-License-Identifier: MPL-2.0
 
-# add_subdirectory(epub)
-# add_subdirectory(ui)
-# add_subdirectory(sd_card)
+add_subdirectory(epub)
+add_subdirectory(sd)
+add_subdirectory(ui)
