--- lib/Kconfig
@@ -1,10 +1,0 @@
-# SPDX-FileCopyrightText: 2025 Anna-Lena Marx <mail@marx.engineer>
-#
-# SPDX-License-Identifier: MPL-2.0
-
-menu "Libraries"
-
-#rsource "ui/Kconfig"
-#rsource "epub/Kconfig"
-
-endmenu
