+++ lib/sd/CMakeLists.txt
@@ -4,6 +4,8 @@
 
 zephyr_library()
 
+zephyr_library_include_directories(. ${ZEPHYR_BASE}/../modules/fs/fatfs/include)
+
 zephyr_library_sources(
-    epub.c
+    sd.c
 )
