+++ lib/ui/CMakeLists.txt
@@ -5,5 +5,6 @@
 zephyr_library()
 
 zephyr_library_sources(
-
+    context.c
+    ui.c
 )
